Damage, Projectile 객체 단순화. Neo Damage policy modified. MechanicLost and MechanicDecay event added. LostAssemblyManager added.

+++ Assets/GO/Cell/Plasm/CellPlasm.cs
@@ -7,7 +7,7 @@ namespace HX
 	{
 		public CellPlasmData data { get; private set; }
 
-		public int hpLeft { get; private set; }
+		public float hpLeft { get; private set; }
 		public Action<CellPlasm> postDecay;
 
 		public DamageDetector damageDetector;
